% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_superagers_network.R
\name{en_superagers_network}
\alias{en_superagers_network}
\title{Elastic net superagers by network}
\usage{
en_superagers_network(
  dat,
  caret = TRUE,
  lambda = "oneSE",
  lambda_grid = NULL,
  alpha = 0.5,
  pmax = 10
)
}
\arguments{
\item{dat}{list of ct_long and sa_long dataframes}

\item{caret}{use package? FALSE is glmnet. They both use caret but have different arguments.}

\item{lambda}{oneSE or best?}

\item{lambda_grid}{Define lambdas in advance; vector}

\item{alpha}{The elasticnet mixing parameter, with \eqn{0≤\alpha≤1}. The penalty is defined as
RVZeat40Jj6yqpBAMFnuX5afqqE1cxD5-\code{alpha=1} is the lasso penalty,
\code{alpha=0} the ridge penalty.}

\item{pmax}{Limit the maximum number of variables ever to be non-zero}
}
\value{
the result of running \link{\code{train} or \link{\code{cv.glmnet}}
}
\description{
Use (tidy) long data format as input.
}
