% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stats.R
\name{fit_stats}
\alias{fit_stats}
\title{Model fit statistics}
\usage{
fit_stats(dat, res, glmnet = FALSE, s = NA)
}
\arguments{
\item{dat}{Data list of control and superager in long format.}

\item{glmnet}{Logical. Which package engine to use.}

\item{s}{lambda; default \code{NA}}
}
\value{
RMSE, R-squared, obs vs predicted
}
\description{
Model fit statistics
}
\references{
\url{https://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}
\url{https://glmnet.stanford.edu/articles/glmnet.html}
}
